+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Allow to translate scope
 * Allow scope based select input values
 * Fix and improve DateTimeInput and DurationInput
+* Translate scope in token selection
 
 ## 0.8.4 - 2019-02-15